Proof of Concept: Test one function using testy
This is just a PoC to show the feasability of doctest:ing the JSDoc
@examples instead of / in addition to other tests.

Because of require/import discrepancies, the test scaffolding needs
some import shenanigans.

Apart from that, the rounding needs to be handled somewhat differently.
This could be solved in other ways, if given a feature request to testy:
- One could imagine the option to inject different expect-logic in the
test suite (make expect behave more fuzzy).
- Or, the option of injecting a helper function to use in @examples.
- ... or other options.

In anticipation of a more thorough effort, though, it seems pertinent
to update and modernize the project first, though.

Nice library! 🙌